modularizing tests checkpoint
