+++ tests/integration_test.py
@@ -1,0 +1,14 @@
+import unittest
+import os
+from pathlib import Path
+
+# Overall integration test
+# (when I fiure out how Python does integration testing!)
+
+class TestRename(unittest.TestCase):
+    
+
+        
+
+if __name__ == '__main__':
+    unittest.main()
